# public/api-doc.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/mine-92b/catway_api/1.0.0
info:
  version: "1.0.0"
  title: Catway-api
  description: The API for the catway management and reservation in Russell harbor.
paths:
  /users:  
    post:
      tags:
        - Users
      description: Create a new user in the database
      operationId: addUser
      security:
        - cookieAuth: []
      parameters:
        - name: token
          in: cookie
          required: true
          description: JWT token in Bearer format
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - lastName
                - firstName
                - email
                - password
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                  description: Must be unique, two users can't have the same email
                password:
                  type: string
              example:
                firstName: "Jean"
                lastName: "Dupont"
                email: "jean-dupont@gmail.com"
                password: "strong-password1234"
      responses:
        '201':
          description: User successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "L'utilisateur Jean Dupont a bien été ajouté"
        '403':
          description: Email already used by a registered user
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Un utilisateur utilise déjà cette adresse mail"
        '400':
          description: An unexpected error occurs
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "L'utilisateur Jean Dupont n'a pas pu être ajouté. Assurez-vous d'avoir rempli tous les champs."
  '/users/{id}':
    get:
      tags:
        - Users
      description: Returns user's data by ID
      operationId: getUser
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: token
          in: cookie
          required: true
          description: JWT token in Bearer format
          schema:
            type: string
      responses:
        '200':
          description: Retrieved user
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                  lastName:
                    type: string
                  firstName:
                    type: string
                  email:
                    type: string
                    format: email
                example:
                  _id: "68375a1dea0fe594ba3afaf1"
                  lastName: "Dupont"
                  firstName: "Jean"
                  email: "jean-dupont@gmail.com"
        '404': 
          description: User not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Cet identifiant ne correspond à aucun utilisateur"
        '501': 
          description: Unexpected Error occurs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse501'
    patch:
      tags:
        - Users
      description: Update user with provided data
      operationId: updateUser
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - lastName
                - firstName
              properties:
                email:
                  type: string
                  format: email
                lastName:
                  type: string
                firstName:
                  type: string
                currentPassword:
                  type: string
                newPassword:
                  type: string
              example:
                email: "john.doe@gmail.com"
                lastName: "Doe"
                firstName: "John"
                currentPassword: "NotEnoughStrongPassword"
                newPassword: "StrongerPassword"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: token
          in: cookie
          required: true
          description: JWT token in Bearer format
          schema:
            type: string
      responses:
        '204':
          description: The user is successfully updated
        '400':
          description: A request body condition is not satisfied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                missingFields:
                  description: Email, last name or first name is/are missing
                  value:
                    message: "Les champs Email, Nom et Prénom sont requis."
                invalidNewPassword:
                  description: New password must contain at least one character
                  value:
                    message: "Le nouveau mot de passe doit contenir au moins un caractère."
        '403':
          description: Invalid current password
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Mot de passe actuel incorrect."
        '404':
          description: User not found with provided ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Cet identifiant ne correspond à aucun utilisateur"
        '501':
          description: An unexpected error occurs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse501'
    delete:
      tags:
        - Users
      description: Delete user by ID
      operationId: deleteUser
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: token
          in: cookie
          required: true
          description: JWT token in Bearer format
          schema:
            type: string
      responses:
        '204':
          description: User is successfully deleted
        '404':
          description: User not found with provided data
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                example:
                  message : "Cet identifiant ne correspond à aucun utilisateur"
        '501':
          description: An unexpected error occurs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse501'
  /users/authenticate:
    post:
      tags:
        - Authenticate
      description: Check authentication by email and password, then send cookies
      operationId: authenticateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              example:
                email: "jean.dupont@gmail.com"
                password: "strongPassword1234"
      responses:
        '200':
          description: User is successfully authenticated
          headers:
            Set-Cookie:
              description: >
                Three cookies in response send :
                - ***token***: Authenticate JWT token, httpOnly, Secured, SameSite=Lax
                - ***firstName***: User first name (non httpOnly)
                - ***lastName***: User last name (non httpOnly)
                
                Example :
                ```
                Set-Cookie: token=eyJhbGciOi...; HttpOnly; Secure; SameSite=Lax; Path=/;
                Set-Cookie: firstName=Jean; Secure; SameSite=Lax; Path=/;
                Set-Cookie: lastName=Dupont; Secure; SameSite=Lax; path=/;
                ```
              schema:
                type: array
                items:
                  type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Vous êtes connectés Jean Dupont."
        '403':
          description: Invalid email and/or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Email et/ou mot de passe incorrect"
        '404':
          description: Invalid email
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Email incorrect"
        '501':
          description: An unexpected error occurs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse501'
  /users/logout:
    post:
      tags:
        - Logout
      description: Delete cookies created during authentication to log the user out
      operationId: logout
      parameters:
          - name: token
            in: cookie
            description: JWT token in Bearer format
            schema:
              type: string
          - name: firstName
            in: cookie
            description: User first name
            schema:
              type: string
          - name: lastName
            in: cookie
            description: User last name
            schema:
              type: string
      responses:
        '200':
          description: User successfully logged out
          headers:
            Set-Cookie:
              description: >
                Cookies deleted :
                - `token` (httpOnly, Secure)
                - `firstName`
                - `lastName`
                
                Example :
                ```
                Set-Cookie: token=; Max-Age=0; Path=/; SameSite=Lax; HttpOnly; Secure
                Set-Cookie: firstName=; Max-Age=0; Path=/; SameSite=Lax; Secure
                Set-Cookie: lastName=; Max-Age=0; Path=/; SameSite=Lax; Secure
                ```
              schema:
                type: array
                items:
                  type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                example:
                  message: "Vous êtes bien déconnecté."
...
components:
  schemas:
    ApiResponse501:
      type: object
      properties:
        message:
          type: string
      example:
        message: "Nous ne parvenons pas à nous connecter à notre base de données. Veuillez réessayer ultérieurement."
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token